Move collections utils to their own module
